move css styles of analysis-tab to its own file
