+++ src/app/tabs/analysis-tab.js
@@ -1,32 +1,4 @@
 var yo = require('yo-yo')
-var remixLib = require('remix-lib')
-
-// -------------- styling ----------------------
-var csjs = require('csjs-inject')
-var styleGuide = remixLib.ui.themeChooser
-var styles = styleGuide.chooser()
-
-var css = csjs`
-  .analysisTabView {
-    padding: 2%;
-    padding-bottom: 3em;
-    display: flex;
-    flex-direction: column;
-  }
-  #staticanalysisView {
-    display: block;
-  }
-  .infoBox  {
-    ${styles.infoTextBox}
-    margin-bottom: 1em;
-  }
-  .textBox  {
-    ${styles.textBoxL}
-    margin-bottom: 1em;
-  }
-`
-
-module.exports = analysisTab
 
 function analysisTab (container, appAPI, events, opts) {
   var el = yo`
@@ -36,3 +8,5 @@ function analysisTab (container, appAPI, events, opts) {
   container.appendChild(el)
   return el
 }
+
+module.exports = analysisTab
